* Passed the part of the path before the last slash instead of the full path of the file * Used *ctypes.c_uint_64 instead of *ctypes.c_int as **char in IMPORT_CALLBACK declaration

+++ c-bindings-tests/compat_test.py
@@ -87,7 +87,7 @@ IMPORT_CALLBACK = ctypes.CFUNCTYPE(
     ctypes.POINTER(ctypes.c_char),
     ctypes.POINTER(ctypes.c_char),
     # we use *int instead of **char to pass the real C allocated pointer, that we have to free
-    ctypes.POINTER(ctypes.c_int),
+    ctypes.POINTER(ctypes.c_uint64),
     ctypes.POINTER(ctypes.c_int)
 )
